+++ scripts/ex-gwf-lak-p02.py
@@ -134,16 +134,16 @@ shape3d = (nlay, nrow, ncol)
 
 data_pth = os.path.join("..", "data", sim_name)
 fpth = os.path.join(data_pth, "idomain-01.txt")
-idomain0 = np.loadtxt(fpth, dtype=np.int)
+idomain0 = np.loadtxt(fpth, dtype=int)
 fpth = os.path.join(data_pth, "idomain-02.txt")
-idomain1 = np.loadtxt(fpth, dtype=np.int)
+idomain1 = np.loadtxt(fpth, dtype=int)
 idomain = [idomain0, idomain1, 1, 1, 1]
 
 # create linearly varying evapotranspiration surface
 
 xlen = delr.sum() - 0.5 * (delr[0] + delr[-1])
 x = 0.0
-s1d = H1 * np.ones(ncol, dtype=np.float)
+s1d = H1 * np.ones(ncol, dtype=float)
 for idx in range(1, ncol):
     x += 0.5 * (delr[idx - 1] + delr[idx])
     frac = x / xlen
@@ -1146,12 +1146,12 @@ def plot_lak_results(gwf, silent=True):
     gwf_results = np.genfromtxt(fpth, delimiter=",", names=True)
 
     dtype = [
-        ("time", np.float),
-        ("LAKE1", np.float),
-        ("LAKE2", np.float),
-        ("A", np.float),
-        ("B", np.float),
-        ("C", np.float),
+        ("time", float),
+        ("LAKE1", float),
+        ("LAKE2", float),
+        ("A", float),
+        ("B", float),
+        ("C", float),
     ]
 
     results = np.zeros((lak_results.shape[0] + 1), dtype=dtype)
